+++ config.go
@@ -10,11 +10,19 @@ type Config struct {
 
 	// Optional. Default: "requestid"
 	RequestIDContextKey string
+
+	// Optional. Default: "info"
+	LogLevel string
+
+	// Optiona. Default: "micro". Possible Value: ["nano", "micro", "milli"]
+	ElapsedTimeUnit string
 }
 
 var ConfigDefault = Config{
 	Next:                nil,
 	RequestIDContextKey: "requestid",
+	LogLevel:            "info",
+	ElapsedTimeUnit:     "micro",
 }
 
 func configDefault(config ...Config) Config {
@@ -29,6 +37,14 @@ func configDefault(config ...Config) Config {
 	// set default
 	if cfg.RequestIDContextKey == "" {
 		cfg.RequestIDContextKey = ConfigDefault.RequestIDContextKey
+	}
+
+	if cfg.LogLevel == "" {
+		cfg.LogLevel = ConfigDefault.LogLevel
+	}
+
+	if cfg.ElapsedTimeUnit == "" {
+		cfg.ElapsedTimeUnit = ConfigDefault.ElapsedTimeUnit
 	}
 
 	return cfg